trade live view and min usdc coin protection

+++ maxhanna.Server/Controllers/DataContracts/Trade/TradebotStatusRequest.cs
@@ -4,7 +4,7 @@ public class TradebotStatusRequest
 	public required int UserId { get; set; }
 	public required string Coin { get; set; }
 	public string? Strategy { get; set; }
-	public int? Hours { get; set; }
+	public double? Hours { get; set; }
 	public int? Page { get; set; }
 	public int? PageSize { get; set; }
 }
